+++ src/users_groups_lib/errors/__init__.py
@@ -1,4 +1,3 @@
 """Exposed errors classes and methods."""
-from users_groups_lib.errors.user import UserExistError
-from users_groups_lib.errors.user import UserPermissionError
+from users_groups_lib.errors.user import UserExistError, UserPermissionError, UserNotExistError
 from users_groups_lib.errors.group import GroupNotExistError