+++ src/users_groups_lib/errors/user/__init__.py
@@ -1,3 +1,4 @@
 """Exposed user errors classes and methods."""
 from users_groups_lib.errors.user.user_exist_error import UserExistError
 from users_groups_lib.errors.user.user_permission_error import UserPermissionError
+from users_groups_lib.errors.user.user_not_exist_error import UserNotExistError